+++ test/bigtest/interactors/patronBlockConditions/settings-conditions-form.js
@@ -2,11 +2,11 @@ import {
   interactor,
   isPresent,
   text,
-  scoped,
-  property
+  property,
+  clickable,
 } from '@bigtest/interactor';
 
-import CheckboxInteractor from '@folio/stripes-components/lib/Checkbox/tests/interactor';
+import CalloutInteractor from '@folio/stripes-components/lib/Callout/tests/interactor';
 
 @interactor class SettingsConditionsForm {
   form = isPresent('[data-test-conditions-form]');
@@ -14,7 +14,12 @@ import CheckboxInteractor from '@folio/stripes-components/lib/Checkbox/tests/int
   blockBorrowing = property('[data-test-block-borrowing]', 'checked');
   blockRenewals = property('[data-test-block-renewals]', 'checked');
   blockRequests = property('[data-test-block-requests]', 'checked');
-  message = scoped('[data-test-block-message]', CheckboxInteractor);
+  blockRequestsClick = clickable('[data-test-block-requests]');
+  message = property('[data-test-block-message]', 'value');
+  saveButton = isPresent('[data-test-charged-out-conditions-save-button]');
+  saveButtonClick = clickable('[data-test-charged-out-conditions-save-button]');
+  isSaveButtonDisabled = property('[data-test-charged-out-conditions-save-button]', 'disabled');
+  calloutMessage = new CalloutInteractor();
 }
 
 export default new SettingsConditionsForm('[class*=conditionsWrapper]');