+++ test/bigtest/network/scenarios/conditions.js
@@ -1,6 +1,0 @@
-/* istanbul ignore file */
-import CQLParser from '../cql';
-
-export default (server) => {
-  server.createList('patronBlockCondition', 6);
-};